Revert "Andy/Initialize Firebase Emulator"
