+++ client/.gitignore
@@ -22,12 +22,3 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
-
-# Firebase Emulator
-.firebaserc
-firebase-debug.log
-firestore-debug.log
-firebase.json
-firestore.indexes.json
-firestore.rules
-storage.rules
