--- client/src/global-state/firebaseEmulator.ts
@@ -1,11 +1,0 @@
-import { connectAuthEmulator } from "firebase/auth";
-import { connectFirestoreEmulator } from "firebase/firestore";
-import { connectStorageEmulator } from "firebase/storage";
-import { auth, storage, db } from "./firebaseSetup";
-
-// use emulator in development
-export const emulator = () => {
-  connectAuthEmulator(auth, "http://localhost:9099");
-  connectFirestoreEmulator(db, "localhost", 8080);
-  connectStorageEmulator(storage, "localhost", 9199);
-};
