+++ client/src/global-state/firebaseSetup.ts
@@ -1,6 +1,5 @@
 import { initializeApp } from "firebase/app";
-import { getAuth } from "firebase/auth";
-import { getStorage } from "firebase/storage";
+import { getAuth, GithubAuthProvider, signInWithPopup } from "firebase/auth";
 import {
   collection,
   getDocs,
@@ -33,16 +32,11 @@ const firebaseConfig = {
   appId: "1:725295282240:web:a82e2b527f44c4c57587a5",
   measurementId: "G-8RVZLGJ0CL",
 };
+import { useAppStore } from "./globalState";
 
 const app = initializeApp(firebaseConfig);
-export const db = getFirestore(app);
-export const auth = getAuth(app);
-export const storage = getStorage(app);
-
-// enable for emulator config
-// import { emulator } from "./firebaseEmulator";
-// emulator();
-
+const db = getFirestore(app);
+const auth = getAuth(app);
 export const usersCol = collection(db, "users");
 export const teamsCol = collection(db, "teams");
 export const projectsCol = collection(db, "projects");
@@ -73,4 +67,13 @@ export async function loadDummyData() {
   loadDataForCol(dummyRepos, reposCol);
   loadDataForCol(dummyProjects, projectsCol);
   loadDataForCol(dummyImplementations, implementationsCol);
+}
+
+export async function githubSignIn() {
+  try {
+    const provider = new GithubAuthProvider();
+    signInWithPopup(auth, provider);
+  } catch (e) {
+    console.log(e);
+  }
 }
